separate core submodule from POJO generation logic fix local dependency integration
