+++ gdk-core/src/main/java/com/github/libgraviton/gdk/api/endpoint/exception/UnableToPersistEndpointAssociationsException.java
@@ -1,4 +1,4 @@
-package com.github.libgraviton.gdk.generator.exception;
+package com.github.libgraviton.gdk.api.endpoint.exception;
 
 public class UnableToPersistEndpointAssociationsException extends Exception {
