+++ gdk-maven-plugin/lib/org/jsonschema2pojo/jsonschema2pojo-core/0.4.28-SNAPSHOT/jsonschema2pojo-core-0.4.28-SNAPSHOT.pom
