+++ gdk-maven-plugin/src/main/java/com/github/libgraviton/gdk/generator/GeneratorInstructionLoader.java
@@ -1,5 +1,7 @@
 package com.github.libgraviton.gdk.generator;
 
+import com.github.libgraviton.gdk.generator.GeneratorInstruction;
+
 import java.util.List;
 
 /**
