+++ gdk-maven-plugin/src/main/java/com/github/libgraviton/gdk/generator/exception/GeneratorException.java
