+++ gdk-maven-plugin/src/main/java/com/github/libgraviton/gdk/generator/instructionloader/grvprofile/EndpointDefinition.java
