+++ gdk-maven-plugin/src/main/java/com/github/libgraviton/gdk/generator/instructionloader/grvprofile/GrvProfileInstructionLoader.java
@@ -7,6 +7,8 @@ import com.github.libgraviton.gdk.exception.CommunicationException;
 import com.github.libgraviton.gdk.exception.SerializationException;
 import com.github.libgraviton.gdk.generator.GeneratorInstruction;
 import com.github.libgraviton.gdk.generator.GeneratorInstructionLoader;
+import com.github.libgraviton.gdk.generator.instructionloader.grvprofile.EndpointDefinition;
+import com.github.libgraviton.gdk.generator.instructionloader.grvprofile.Service;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -22,7 +24,7 @@ import java.util.List;
  */
 public class GrvProfileInstructionLoader implements GeneratorInstructionLoader {
 
-    private final Logger LOG = LoggerFactory.getLogger(GrvProfileInstructionLoader.class);
+    private final Logger LOG = LoggerFactory.getLogger(com.github.libgraviton.gdk.generator.instructionloader.grvprofile.GrvProfileInstructionLoader.class);
 
     /**
      * The GravitonApi instance where the endpoint definitions will be loaded from.
