+++ gdk-maven-plugin/src/main/java/com/github/libgraviton/gdk/generator/instructionloader/grvprofile/Service.java
@@ -3,6 +3,7 @@ package com.github.libgraviton.gdk.generator.instructionloader.grvprofile;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.github.libgraviton.gdk.generator.instructionloader.grvprofile.EndpointDefinition;
 
 import java.util.List;
