+++ gdk-maven-plugin/src/test/resources/service/grvProfileInstructionLoaderTest.json
