+++ gdk-maven-plugin/src/test/resources/service/serviceTest.json
