+++ gdk-maven-plugin/src/test/resources/serviceSchema/grvProfileInstructionLoaderTest.anotherMoreSchema.json
