+++ gdk-maven-plugin/src/test/resources/serviceSchema/grvProfileInstructionLoaderTest.anotherSchema.json
