+++ gdk-maven-plugin/src/test/resources/serviceSchema/grvProfileInstructionLoaderTest.someMoreSchema.json
