+++ gdk-maven-plugin/src/test/resources/serviceSchema/grvProfileInstructionLoaderTest.someSchema.json
